requirements.txt
opencv-python
numpy
ultralytics
reolinkapi
python-dotenv
flask
